Added methods for recurring jobs.

+++ lib/delayed/job.rb
@@ -22,7 +22,7 @@ module Delayed
 
     self.destroy_failed_jobs = true
 
-    NextTaskSQL  = '(`locked_by` = ?) OR (`run_at` <= ? AND (`locked_at` IS NULL OR `locked_at` < ?))'
+    NextTaskSQL  = '(`locked_by` = ? AND `run_at` <= ?) OR (`run_at` <= ? AND (`locked_at` IS NULL OR `locked_at` < ?))'
     NextTaskOrder = 'priority DESC, run_at ASC'
     ParseObjectFromYaml = /\!ruby\/\w+\:([^\s]+)/
 
@@ -74,12 +74,20 @@ module Delayed
 
       Job.create(:payload_object => object, :priority => priority.to_i)
     end
-
+    
+    #This provides a method of "scheduling"
+    #start_time is when the job will first run, the period dictates when it will run again after that
+    def self.recurring(object, priority = 0, start_time = db_time_now, period = 24.hours)
+      unless object.respond_to?(:perform)
+        raise ArgumentError, 'Cannot enqueue items which do not respond to perform'
+      end
+      Job.create(:payload_object => object, :priority => priority.to_i, :run_at => start_time, :recur => true, :period => period)    
+    end
      
     def self.find_available(limit = 5, max_run_time = MAX_RUN_TIME)
       time_now = db_time_now      
       sql = NextTaskSQL.dup
-      conditions = [time_now, time_now + max_run_time, worker_name]
+      conditions = [time_now, time_now, time_now, worker_name]
       
       if self.min_priority
         sql << ' AND (`priority` >= ?)'
@@ -90,6 +98,7 @@ module Delayed
         sql << ' AND (`priority` <= ?)'
         conditions << max_priority         
       end
+      
            
       conditions.unshift(sql)         
 
@@ -113,7 +122,12 @@ module Delayed
           job.lock_exclusively!(max_run_time, worker_name)
           runtime =  Benchmark.realtime do
             yield job.payload_object
-            job.destroy
+            if job.recur == true
+              run = job.run_at + job.period
+              job.update_attributes(:run_at => run)
+            else
+              job.destroy
+            end
           end
           logger.info "* [JOB] #{job.name} #{job.description}completed after %.4f" % runtime
           
@@ -178,7 +192,7 @@ module Delayed
 
         job = self.reserve do |j|
           begin
-            logger.info "job: #{j.description}"
+            logger.info "job: #{j.description}" if j.description
             j.perform
             success += 1
           rescue
